WIP start migrating AnalysisHandler
