+++ raml/schemas/mongo/analysis.json
@@ -4,6 +4,13 @@
     "type": "object",
     "properties": {
     	"_id":          {"type": "string"},
+        "parent":       {
+                            "type": "object",
+                            "properties": {
+                                "type": "string",
+                                "id":   "string"
+                            }
+                        },
         "created":      {},
         "modified":     {},
         "notes":        {"type": "array",  "items": {"$ref": "note.json"}},
